loader/requirements.txt
# using code from https://github.com/geopandas/geopandas/pull/1248
-e git+https://github.com/HTenkanen/geopandas#egg=geopandas
SQLAlchemy
psycopg2-binary
geoalchemy2
datapackage
intake
intake-parquet
intake_geopandas
click
prefect
